Fix traceback messages when script cancelled by user.
This fixes #24. When user cancelled the scripts, an error was being
showed. The correct behavior would be to print a informative
message and exit gracefully.

When cancelled, ideally, the prompts should raise a KeyboardInterrupt.
However, they do not raise exceptions, but show messages and exit
return None. We decided to just catch the exceptions, incurring in
the risk of catching other exceptions that should be dealt properly.
The other option would be to catch exceptions on every prompt. At
the momment, the payback does not worth the work.